radicals_bot/pyproject.toml
[tool.poetry]
name = "radicals-bot"
version = "0.1.0"
description = ""
authors = ["Slmpnv <slmpnv.code.mail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.15.4"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
icecream = "^2.1.3"
path = "^16.10.0"
matplotlib = "^3.8.2"
sdamgia-api = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
